+++ src/constants.ts
@@ -1,1 +1,3 @@
 export const UNIRTOS_REPO = 'https://github.com/UniRTOS/unirtos.git';
+export const UNIRTOS_VENV = 'unirtos-venv';
+export const CONFIG_FILE = 'env_config.json';